move resources in a module
